+++ .vscode/cSpell.json
@@ -7,6 +7,7 @@
     // words - list of words to be always considered correct
     "words": [
         "Flexbox",
+        "doesnt",
         "filepath",
         "folderpath",
         "namepath",
