+++ storybook/shared/storyFolders/index.js
@@ -1,0 +1,15 @@
+/* eslint-disable global-require, import/no-dynamic-require */
+
+const { join } = require('path')
+const { loadTest } = require('./test')
+const { loadWebpack } = require('./webpack')
+
+const isTest =
+  process.env.NODE_ENV === 'test' || process.env.NODE_ENV === 'test:image'
+const path = '../../stories'
+
+const storyFolders = !isTest
+  ? loadWebpack(require.context(path, true, /.+\.(md|png|js)$/i))
+  : loadTest(join(__dirname, path))
+
+module.exports = { storyFolders }
